create and style the exams index view
this was 100% done by cursor. i simply asked it this: "write the index view for the exams controller, making it similar to the index view of the lessons controller. use the same design and style, similar cards and structure. be sure to use the proper attributes for each exam tho"

this is insane, completely insane